# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/oauth2/clients": {
            "get": {
                "description": "分页查询所有 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "分页获取 OAuth2 客户端配置列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认20）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "为指定邮箱提供商创建新的 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "创建 OAuth2 客户端配置",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/clients/provider/{provider_name}/default": {
            "get": {
                "description": "获取指定邮箱提供商的默认 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "获取指定提供商的默认客户端配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱提供商类型 (1=Gmail, 2=Outlook)",
                        "name": "provider_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/clients/provider/{provider_type}": {
            "get": {
                "description": "获取指定邮箱提供商的所有 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "获取指定提供商的所有客户端配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱提供商类型 (1=Gmail, 2=Outlook)",
                        "name": "provider_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/clients/smart-select": {
            "get": {
                "description": "智能选择 OAuth2 客户端，支持配额管理和自动降级",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "智能选择客户端配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱提供商类型 (1=Gmail, 2=Outlook)",
                        "name": "provider_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "指定的客户端 ID（可选）",
                        "name": "client_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/clients/{id}": {
            "get": {
                "description": "根据 ID 获取 OAuth2 客户端配置详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "获取单个 OAuth2 客户端配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "客户端 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定 ID 的 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "更新 OAuth2 客户端配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "客户端 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fusionmail_internal_model.OAuth2ClientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定 ID 的 OAuth2 客户端配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "删除 OAuth2 客户端配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "客户端 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth2/clients/{id}/default": {
            "post": {
                "description": "将指定客户端设置为提供商的默认配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2客户端管理"
                ],
                "summary": "设置默认客户端配置",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "客户端 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱提供商类型 (1=Gmail, 2=Outlook)",
                        "name": "provider_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fusionmail_internal_model.OAuth2ClientCreateRequest": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret",
                "name",
                "provider_id",
                "redirect_uri"
            ],
            "properties": {
                "client_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "client_secret": {
                    "type": "string"
                },
                "metadata": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "provider_id": {
                    "description": "提供商ID",
                    "type": "integer"
                },
                "quota_daily": {
                    "type": "integer",
                    "minimum": -1
                },
                "quota_monthly": {
                    "type": "integer",
                    "minimum": -1
                },
                "redirect_uri": {
                    "type": "string"
                }
            }
        },
        "fusionmail_internal_model.OAuth2ClientResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "last_used_at": {
                    "type": "string"
                },
                "metadata": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "integer"
                },
                "provider_name": {
                    "description": "提供商名称（通过关联获取）",
                    "type": "string"
                },
                "quota_daily": {
                    "type": "integer"
                },
                "quota_monthly": {
                    "type": "integer"
                },
                "redirect_uri": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "type": "integer"
                }
            }
        },
        "fusionmail_internal_model.OAuth2ClientUpdateRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "client_secret": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "metadata": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "provider_id": {
                    "description": "提供商ID（可选）",
                    "type": "integer"
                },
                "quota_daily": {
                    "type": "integer",
                    "minimum": -1
                },
                "quota_monthly": {
                    "type": "integer",
                    "minimum": -1
                },
                "redirect_uri": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}